requirement.txt
tensorflow==2.0
gensim
anyconfig
opencv-python
pillow